Movement update
Tried to fix and keep the original movement code since it's a lot smaller and cleaner then the one I had to make. But was causing to many issues.
I believe at the end I noticed a way to keep it and will try again at a later date.

For now I've implemented a none buggy working movement script and added in the basics of player interaction UI