Added Navpoints and event system to gate controls
Added Navpoints to all areas of the map. Any navpoints for guards in the Hubs are contained in an empty object in the hubs. Any not in a hubs are contained in the global NavPotions empty object.

Also added missing control panels to the gates and attached their event objects.